requirements.txt
Pillow==8.4.0
torchvision==0.12.0
mpmath==1.2.1
numpy==1.21.4
multitasking==0.0.10
pandas==1.3.4
pandas-datareader==0.10.0
pyobjc-framework-AutomaticAssessmentConfiguration==8.4.1
openpyxl==3.0.9
pyobjc-framework-ExecutionPolicy==8.4.1
pyobjc-framework-InputMethodKit==8.4.1
svgpathtools==1.4.4
multitasking==0.0.10
pandas==1.3.4
pandas-datareader==0.10.0
pyobjc-framework-AutomaticAssessmentConfiguration==8.4.1
CairoSVG==2.5.2
svglib==1.2.1
svgpathtools==1.4.4
svgwrite==1.4.2
torch==1.11.0
torchaudio==0.11.0
torchfile==0.1.0
torchvision==0.12.0
multitasking==0.0.10
pandas==1.3.4
pandas-datareader==0.10.0